Added new tests

+++ multithread/mtp-lab3/data-handler/src/algorithm.c
@@ -23,8 +23,8 @@ void swap(uint64_t *A, uint64_t *B) {
 }
 
 void bubbleSort(uint64_t *array, long size) {
-    for (int k = 0; k < size - 1; k++)
-        for (int i = 0; i < size - k - 1; i++)
+    for (long k = 0; k < size - 1; k++)
+        for (long i = 0; i < size - k - 1; i++)
             if (array[i] > array[i + 1])
                 swap(&array[i], &array[i + 1]);
 }
